Refactor Julia GC integration tests.
This introduces a new make target, testkernel.
