+++ src/julia_gc.c
@@ -997,116 +997,3 @@ void MarkJuliaWeakRef(void * p)
     if (JMarkTyped(p, jl_weakref_type))
         YoungRef++;
 }
-
-
-#ifdef TEST_JULIA_GC_INTERNALS
-
-static int int_cmp(int a, int b)
-{
-    return a - b;
-}
-
-#define ELEM_TYPE int
-#define COMPARE int_cmp
-
-#include "baltree.h"
-
-void TestScapegoatTrees(void)
-{
-    const Int N = 1024 * 1024;
-    intTree * btree = intTreeMake();
-    for (int i = 0; i < N; i++) {
-        intTreeInsert(btree, i);
-    }
-    if (intTreeCount(btree) != N)
-        abort();
-    if (height_to_size[intTreeDepth(btree) - 1] > intTreeCount(btree))
-        abort();
-    for (int i = 0; i < N; i++) {
-        if (!intTreeFind(btree, i))
-            abort();
-    }
-    for (int i = 1; i < N; i += 2) {
-        intTreeRemove(btree, i);
-    }
-    for (int i = 0; i < N; i += 2) {
-        if (!intTreeFind(btree, i))
-            abort();
-        if (intTreeFind(btree, i + 1))
-            abort();
-    }
-    for (int i = 1; i < N; i++) {
-        intTreeRemove(btree, i);
-    }
-    if (intTreeCount(btree) != 1 || intTreeDepth(btree) != 1)
-        abort();
-    intTreeClear(btree);
-    for (int i = 0, j = 0; i < N; i++, j = (5 * j + 1) & (N - 1)) {
-        intTreeInsert(btree, j);
-    }
-    if (intTreeCount(btree) != N)
-        abort();
-    if (height_to_size[intTreeDepth(btree) - 1] > intTreeCount(btree))
-        abort();
-    for (int i = 0; i < N; i++) {
-        if (!intTreeFind(btree, i))
-            abort();
-    }
-    for (int i = 0, j = 0; i < N; i++, j = (j + 7) & (N - 1)) {
-        if (i == N / 2 && intTreeCount(btree) != N / 2)
-            abort();
-        intTreeRemove(btree, i);
-    }
-    if (intTreeCount(btree) != 0 || intTreeDepth(btree) != 0)
-        abort();
-    intTreeDelete(btree);
-    printf("Scapegoat tree tests passed.\n");
-}
-
-#define ELEM_TYPE int
-#define COMPARE int_cmp
-
-#include "dynarray.h"
-
-static void TestDynArrays(void)
-{
-    const Int  N = 1024 * 1024;
-    intArray * arr = intArrayMake(1);
-    for (Int i = 0; i < N; i++) {
-        intArrayAdd(arr, N - i);
-    }
-    for (Int i = 0; i < N; i++) {
-        if (intArrayGet(arr, i) != N - i)
-            abort();
-    }
-    intArraySort(arr);
-    for (Int i = 0; i < N; i++) {
-        if (intArrayGet(arr, i) != i + 1)
-            abort();
-    }
-    intArraySetLen(arr, N / 2);
-    for (Int i = 0; i < N / 2; i++) {
-        intArrayAdd(arr, arr->items[i]);
-    }
-    if (intArrayLen(arr) != N)
-        abort();
-    intArray * arr2 = intArrayClone(arr);
-    for (Int i = 0; i < N / 2; i++) {
-        if (intArrayGet(arr2, i) != i + 1)
-            abort();
-        if (intArrayGet(arr2, i + N / 2) != i + 1)
-            abort();
-    }
-    intArrayDelete(arr);
-    intArrayDelete(arr2);
-    printf("Dynamic array tests passed.\n");
-}
-
-__attribute__((constructor)) static void RunTests(void)
-{
-    TestScapegoatTrees();
-    TestDynArrays();
-    exit(0);
-}
-
-#endif
